use queue as log callback to avoid unsafe calls from trap context
The LogCallback is sometimes also triggered from trap contexts. Trap
may only run a subset of Ruby to avoid deadlocks. However, mutex
locks are not safe (see [1]) and the standard logger uses a mutex
to prevent logs from interleaving [2]. In this scenario it would
print "log writing failed.  can't be called from trap context"
(via [3], [4]) on stderr, but omit the actual error message.

This patch introduces a queue which is safe to call from traps for
(see [5]) and spawns a separate thread to empty/print this queue.

[1] https://bugs.ruby-lang.org/issues/7917#note-1
[2] https://github.com/ruby/ruby/blob/3816622fbedd034d338fcb1bbdc80d163e302ae6/lib/logger/log_device.rb#L31
[3] https://github.com/ruby/ruby/blob/3816622fbedd034d338fcb1bbdc80d163e302ae6/lib/logger/log_device.rb#L46
[4] https://github.com/ruby/ruby/blob/3816622fbedd034d338fcb1bbdc80d163e302ae6/thread_sync.c#L246
[5] https://bugs.ruby-lang.org/issues/14222#note-3