adapt for additional tests
This changes the tests slightly in the sense that they don't assert on
the logger's formatting anymore. The general formatting is still checked
in the new config_spec test for the queue. Changing it everywhere is
possible, but carries the downside that all tests need to wait for the
log thread to process the messages.