Use a finalizer bound to the object that will be finalized
Previous object finalizer was bound to a string (`@id`) being created on
initialization instead of the object itself.
This aimed to remove a warning about object finalizer reference after
reading [1] but changed the first argument instead of the second that is
mentioned on [1]: `proc { close }` has binding to the own object.

The current implementation removes the warning but still does not work
to run the finalizer when the reference to the producer is removed (like
`producer = nil`) therefore running only on exit. This behavior is
explained on [2].

Ref:
  1. https://www.mikeperham.com/2010/02/24/the-trouble-with-ruby-finalizers/
  2. https://ruby-doc.org/core-3.0.2/ObjectSpace.html#method-c-define_finalizer

---

Changelog: fix
Ref: https://www.mikeperham.com/2010/02/24/the-trouble-with-ruby-finalizers/
Ref: https://ruby-doc.org/core-3.0.2/ObjectSpace.html#method-c-define_finalizer
See-also: https://www.mikeperham.com/2010/02/24/the-trouble-with-ruby-finalizers/
See-also: https://ruby-doc.org/core-3.0.2/ObjectSpace.html#method-c-define_finalizer