Use brew --prefix [formula] to find openssl on a mac
brew --prefix has a form that returns a path to a specific
formula instead of the general brew installation path.

This method of resolving paths is preferred because homebrew started
to add suffixes to openssl installation to distinguish 1.1 vs 3.0,
so openssl started to be installed into `/opt/homebrew/opt/openssl@3`
and `/opt/homebrew/opt/openssl@1.1` instead of the expected
`/ops/homebrew/opt/openssl`.